+++ backend/core/views.py
@@ -1,7 +1,17 @@
+# Third-party
+from jklib.dj.viewsets import ImprovedViewSet
+
 # Django
 from django.http import HttpRequest, HttpResponse
 from django.shortcuts import render
 from django.views.decorators.http import require_GET
+from rest_framework import permissions
+from rest_framework.decorators import action
+from rest_framework.request import Request
+from rest_framework.response import Response
+
+# Application
+from core.serializers import AppConfigSerializer
 
 
 def index(request: HttpRequest) -> HttpResponse:
@@ -20,3 +30,15 @@ def robots_txt(request: HttpRequest) -> HttpResponse:
 @require_GET
 def ping(request: HttpRequest) -> HttpResponse:
     return HttpResponse("pong", content_type="text/plain")
+
+
+class AppViewSet(ImprovedViewSet):
+    default_permission_classes = [permissions.IsAuthenticated]
+    serializer_class_per_action = {
+        "config": AppConfigSerializer,
+    }
+
+    @action(detail=False, methods=["GET"])
+    def config(self, request: Request) -> Response:
+        serializer = self.get_serializer()
+        return Response(serializer.data)