+++ backend/django_react_starter/settings/test.py
@@ -3,7 +3,7 @@ import os
 
 # Local
 from .base import *  # noqa type: ignore
-from .base import BASE_DIR  # noqa type: ignore
+from .base import BASE_DIR, MEDIA_ROOT  # noqa type: ignore
 
 DEBUG = False
 ENVIRONMENT = "test"
@@ -13,6 +13,9 @@ ALLOWED_HOSTS = ["localhost", "127.0.0.1"]
 SECRET_KEY = "django-insecure-cb7(pt2v*1*=%kgz(e(vm9q+wlm4olyz_!(bef$3nlwp)xi)g*"
 DJANGO_SUPERUSER_EMAIL = "random-email@for-test.com"
 DJANGO_SUPERUSER_PASSWORD = "random-password-for-test"
+
+# Media
+MEDIA_ROOT = MEDIA_ROOT + "-test"
 
 # Database
 DATABASES = {