+++ backend/user/tests/test_views.py
@@ -1,31 +1,29 @@
 # Django
 from django.contrib.auth import get_user_model
+from rest_framework.reverse import reverse
 
 # Application
 from core.tests import BaseActionTestCase
-from django_react_starter.urls import API_V1
 from user.tests.factories import UserFactory
 
 User = get_user_model()
 
 
 class AuthViewSetTestCase(BaseActionTestCase):
-    viewset_url = f"/{API_V1}auth/"
-
     def test_check_success(self) -> None:
-        url = self.build_url(extra_path="check")
+        url = reverse("auth-check")
         response = self.api_client.get(url)
         self.assertEqual(response.status_code, 204)
 
     def test_check_error(self) -> None:
-        url = self.build_url(extra_path="check")
+        url = reverse("auth-check")
         self.api_client.logout()
         response = self.api_client.get(url)
         self.assertEqual(response.status_code, 401)
 
     def test_login_success(self) -> None:
         self.api_client.logout()
-        url = self.build_url(extra_path="login")
+        url = reverse("auth-login")
         user = UserFactory(email="test@test.test", password="test")
         payload = {"email": user.email, "password": "test"}
         response = self.api_client.post(url, data=payload)
@@ -35,7 +33,7 @@ class AuthViewSetTestCase(BaseActionTestCase):
 
     def test_login_error(self) -> None:
         self.api_client.logout()
-        url = self.build_url(extra_path="login")
+        url = reverse("auth-login")
         user = UserFactory(email="test@test.test", password="test")
         # Invalid password
         payload = {"email": user.email, "password": "invalid password"}
@@ -54,14 +52,12 @@ class AuthViewSetTestCase(BaseActionTestCase):
 
     def test_logout_success(self) -> None:
         self.assertTrue(self.user.is_authenticated)
-        url = self.build_url(extra_path="logout")
+        url = reverse("auth-logout")
         response = self.api_client.post(url)
         self.assertEqual(response.status_code, 204)
 
 
 class CurrentUserViewSetTestCase(BaseActionTestCase):
-    viewset_url = f"/{API_V1}self/"
-
     def test_create_success(self) -> None:
         self.user.profile.subscribed_to_notifications = False
         self.user.profile.save()
@@ -70,7 +66,8 @@ class CurrentUserViewSetTestCase(BaseActionTestCase):
             "last_name": "Doe",
             "email": "johndoe@johndoe.com",
         }
-        response = self.api_client.post(self.viewset_url, data=payload)
+        url = reverse("self-list")
+        response = self.api_client.post(url, data=payload)
         self.assertEqual(response.status_code, 200)
         self.user.refresh_from_db()
         self.assertEqual(self.user.pk, response.data["id"])
@@ -79,12 +76,13 @@ class CurrentUserViewSetTestCase(BaseActionTestCase):
         self.assertEqual(self.user.email, payload["email"])
 
     def test_list_success(self) -> None:
-        response = self.api_client.get(self.viewset_url)
+        url = reverse("self-list")
+        response = self.api_client.get(url)
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.data["id"], self.user.pk)
 
     def test_update_password_error(self) -> None:
-        url = self.build_url(extra_path="update_password")
+        url = reverse("self-update-password")
         current_password = "stR0ngP4ssw0rd!"
         self.user.set_password(current_password)
         self.user.save()
@@ -115,7 +113,7 @@ class CurrentUserViewSetTestCase(BaseActionTestCase):
         self.assertEqual(response.status_code, 400)
 
     def test_update_password_success(self) -> None:
-        url = self.build_url(extra_path="update_password")
+        url = reverse("self-update-password")
         current_password = "stR0ngP4ssw0rd!"
         self.user.set_password(current_password)
         self.user.save()