+++ frontend/src/core/proptypes/index.js
@@ -1,2 +1,3 @@
+export { AppConfigPropShape } from './appConfig';
 export { CurrentUserPropShape } from './currentUser';
 export { DayjsPropType, RoutePropShape, RouteConfigPropShape } from './custom';